pdf-ocr/requirements.txt
numpy==2.2.4
pytesseract==0.3.13
pdf2image==1.17.0
tqdm==4.67.1
google-generativeai==0.8.4
dotenv==0.9.9